+++ src/main/java/it/gov/pagopa/fdr/controller/middleware/filter/NoInternalSchemaFilter.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.util;
+package it.gov.pagopa.fdr.controller.middleware.filter;
 
 import java.util.Map;
 import java.util.Map.Entry;
