+++ src/main/java/it/gov/pagopa/fdr/util/constant/MDCKeys.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.util;
+package it.gov.pagopa.fdr.util.constant;
 
 public class MDCKeys {
