+++ src/main/java/it/gov/pagopa/fdr/util/re/AppReUtil.java
@@ -1,5 +1,6 @@
-package it.gov.pagopa.fdr.util;
+package it.gov.pagopa.fdr.util.re;
 
+import it.gov.pagopa.fdr.controller.interfaces.annotation.Re;
 import it.gov.pagopa.fdr.service.model.re.FdrActionEnum;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.InvocationTargetException;
