+++ src/main/resources/application.properties
@@ -89,11 +89,11 @@ api_config_cache.cron.expr=${ADAPTER_API_CONFIG_CACHE_CRON:*/15 * * * * ?}
 ###################
 #operations.filter=".*"
 %openapi_internal.operations.filter=/internal/.*
-%openapi_internal.mp.openapi.filter=it.gov.pagopa.fdr.util.NoInternalSchemaFilter
+%openapi_internal.mp.openapi.filter=it.gov.pagopa.fdr.controller.middleware.filter.NoInternalSchemaFilter
 %openapi_psp.operations.filter=/(info|psps/).*
-%openapi_psp.mp.openapi.filter=it.gov.pagopa.fdr.util.NoInternalSchemaFilter
+%openapi_psp.mp.openapi.filter=it.gov.pagopa.fdr.controller.middleware.filter.NoInternalSchemaFilter
 %openapi_organization.operations.filter=/(info|organizations/).*
-%openapi_organization.mp.openapi.filter=it.gov.pagopa.fdr.util.NoInternalSchemaFilter
+%openapi_organization.mp.openapi.filter=it.gov.pagopa.fdr.controller.middleware.filter.NoInternalSchemaFilter
 ##############
 # MOCK SERVER
 ##############
